Nova Versão já cadastrando, Excluindo e Listando.

+++ src/br/aeso/exercicio/cliente/IRepositorioCliente.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 public interface IRepositorioCliente {
 	public void cadastrar(Cliente cliente) throws IOException;
 	public void atualizar(Cliente cliente) throws IOException;
-	public boolean remover(double codigo) throws IOException;
+	public boolean remover(int codigo) throws IOException;
 	public Cliente procurar(double codigo) throws SQLException;
 	public ArrayList<Cliente> listar();
 }
