+++ src/br/aeso/exercicio/produto/IRepositorioProduto.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 public interface IRepositorioProduto {
 	public void cadastrar(Produto produto) throws IOException;
 	public void atualizar(Produto produto) throws IOException;
-	public boolean remover(double codigo) throws IOException;
+	public boolean remover(int codigo) throws IOException;
 	public Produto procurar(double codigo);
 	public ArrayList<Produto> listar();
 }
